+++ CHANGELOG.md
@@ -6,6 +6,8 @@ New in 1.10.0
 * `featureColorForRegex` can set colour for features NOT matching a regex. Useful to dim features without 
 completely hiding them.
 
+* `awk` recognises column headers for bam, vcf, gtf/gff, bed tracks, like `$POS $START $FEATURE` etc, similar to [bioawk](https://github.com/lh3/bioawk).
+
 * A pair of single quotes (`''`) at the command prompt is understood as an empty argument.